requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.29.14
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.2.38
entrypoints==0.3
enum34==1.1.6
gast==0.3.3
google-auth==1.9.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.16.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
PyJWT==1.7.1
pymongo==3.7.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
sqlparse==0.2.4
tb-nightly==2.3.0a20200721
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow @ https://github.com/lakshayg/tensorflow-build-archived/releases/download/old-tensorflow-versions/tensorflow-1.4.0-cp36-cp36m-macosx_10_12_x86_64.whl
tensorflow-estimator==2.0.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.0.0.dev2019122909
tf-nightly==2.4.0.dev20200721
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
